Policy engine test automation (#784)
General test information is available on test/integration/acceptance/custom/hello_policy/README.md; implementation details available on test/integration/acceptance/custom/hello_policy/doc.go.

+++ test/utils/base/cluster_context.go
@@ -13,7 +13,9 @@ import (
 	"github.com/skupperproject/skupper/pkg/kube"
 	"github.com/skupperproject/skupper/test/utils/k8s"
 	apiv1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/client-go/util/retry"
 )
@@ -47,6 +49,27 @@ func (cc *ClusterContext) KubectlExec(command string) ([]byte, error) {
 }
 
 func (cc *ClusterContext) CreateNamespace() error {
+	if ShouldSkipNamespaceSetup() {
+		log.Printf("Skipping namespace creation for %v", cc.Namespace)
+		ns, err := cc.VanClient.KubeClient.CoreV1().Namespaces().Get(cc.Namespace, metav1.GetOptions{})
+		if err == nil {
+			if ns != nil {
+				// As we're skipping the creation of namespaces, we're adopting whatever
+				// we find; we will destroy these when DeleteNamespace is called, unless
+				// ShouldSkipNamespaceTeardown returns true.
+				log.Printf("Reusing existing namespace %v", cc.Namespace)
+				cc.nsCreated = true
+			} else {
+				// Assertion; this should never happen
+				return fmt.Errorf("Namespace check returned nil response, but no errors")
+			}
+		} else {
+			if errors.IsNotFound(err) {
+				return fmt.Errorf("Namespace %v did not exist and namespace creation skipping was requested", cc.Namespace)
+			}
+			return err
+		}
+	}
 	_, err := kube.NewNamespace(cc.Namespace, cc.VanClient.KubeClient)
 	if err == nil {
 		cc.nsCreated = true
@@ -59,12 +82,26 @@ func (cc *ClusterContext) DeleteNamespace() error {
 		log.Printf("namespace [%s] will not be deleted as it was not created by ClusterContext", cc.Namespace)
 		return nil
 	}
+	if ShouldSkipNamespaceTeardown() {
+		log.Print("Skipping namespace tear down, per env variables")
+		return nil
+	}
 	if err := k8s.DeleteNamespaceAndWait(cc.VanClient.KubeClient, cc.Namespace); err != nil {
 		return err
 	}
 
 	cc.nsCreated = false
 	return nil
+}
+
+// As the name says, it will add label to this namespace
+func (cc *ClusterContext) LabelNamespace(label string, value string) (err error) {
+
+	payload := fmt.Sprintf(`{"metadata": {"labels": {"%v": "%v"}}}`, label, value)
+
+	_, err = cc.VanClient.KubeClient.CoreV1().Namespaces().Patch(cc.Namespace, types.MergePatchType, []byte(payload))
+
+	return
 }
 
 func (cc *ClusterContext) waitForSkupperServiceToBeCreated(name string, retryFn func() (*apiv1.Service, error), backoff wait.Backoff) (*apiv1.Service, error) {
